pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "justrayzist"
version = "0.1.0"
description = "Offline local image generation app targeting Z-Image Turbo."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "accelerate>=1.2.0",
  "diffusers>=0.35.0",
  "fastapi>=0.116.0",
  "gguf>=0.17.0",
  "huggingface-hub>=0.35.0",
  "pillow>=11.0.0",
  "psutil>=6.0.0",
  "pyyaml>=6.0.0",
  "safetensors>=0.6.0",
  "torch>=2.9.1",
  "torchaudio>=2.9.1",
  "torchvision>=0.24.1",
  "transformers>=4.56.0",
  "typer>=0.16.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[project.scripts]
justrayzist = "app.cli.main:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
